requirements.txt
scholarly
numpy
pillow
RPi.GPIO